refactor(server-rs): drop sentry-native — the engine links no Sentry SDK
The Rust sentry crate + minidump reporter (already wired) are the only
crash/error-reporting path now; the C++ engine no longer needs one of its
own. Removes sentry-native (Bazel dep, init/close, the terminate handler,
the dead CLI11 --sentry-* flags) and re-homes ImageContext/populate_from_image
into a plain populate_from_image.h with no telemetry dependency.

Two things the plan got wrong, caught by reading the actual code before
editing it: SipiHttpServer.cpp had no live capture_image_error call sites to
convert to log-only (just leftover dead imports), and the CLI path
(cli_app.cpp, convert_access_file.cpp) needed real edits the plan hadn't
scoped — every catch site already calls log_err/emit_json_report alongside
capture_image_error, so removing the latter is a clean, verified no-op for
CLI error visibility.

Part of DEV-6659.
