--- bazel/sentry.BUILD.bazel
@@ -1,109 +1,0 @@
-"""Native `cc_library` for sentry-native 0.9.0.
-
-Source selection mirrors sentry's `src/CMakeLists.txt` for a non-Windows,
-non-musl (glibc-Linux / macOS) build with the **inproc** backend, the **curl**
-transport, and the **libbacktrace** unwinder (`SENTRY_BACKEND=inproc`,
-transport/unwinder left at their unix defaults). Windows / Android / AIX / musl
-paths and the breakpad/crashpad backends are out of scope.
-
-No generated config header: sentry-native is configured purely by compile
-definitions. The four below are the ones its sources actually consume —
-`SENTRY_BACKEND_INPROC` (sentry_scope.c), `SENTRY_WITH_UNWINDER_LIBBACKTRACE`
-(sentry_unwinder.c dispatcher), `SENTRY_HANDLER_STACK_SIZE` (sentry_os.c +
-inproc backend), `SIZEOF_LONG` (sentry_sync.h). `SENTRY_SDK_{NAME,VERSION,
-USER_AGENT}` come from the bundled `include/sentry.h`; `SENTRY_BUILD_STATIC` is
-a no-op on non-Windows (SENTRY_API expands empty without `SENTRY_BUILD_SHARED`).
-
-The unwinder uses `backtrace()` from `<execinfo.h>` — glibc (libc) and macOS
-(libSystem) both provide it; the Linux `-ldl`/`-lpthread` link is declared by
-the //src/metadata and //src:sipi_lib consumers.
-"""
-
-load("@rules_cc//cc:cc_library.bzl", "cc_library")
-
-package(default_visibility = ["//visibility:public"])
-
-_SENTRY_SRCS = [
-    # core (always compiled)
-    "src/sentry_alloc.c",
-    "src/sentry_backend.c",
-    "src/sentry_core.c",
-    "src/sentry_database.c",
-    "src/sentry_envelope.c",
-    "src/sentry_info.c",
-    "src/sentry_json.c",
-    "src/sentry_logger.c",
-    "src/sentry_options.c",
-    "src/sentry_os.c",
-    "src/sentry_ratelimiter.c",
-    "src/sentry_scope.c",
-    "src/sentry_session.c",
-    "src/sentry_slice.c",
-    "src/sentry_string.c",
-    "src/sentry_sync.c",
-    "src/sentry_transport.c",
-    "src/sentry_utils.c",
-    "src/sentry_uuid.c",
-    "src/sentry_value.c",
-    "src/sentry_tracing.c",
-    "src/path/sentry_path.c",
-    "src/screenshot/sentry_screenshot.c",
-    "src/transports/sentry_disk_transport.c",
-    "src/transports/sentry_function_transport.c",
-    "src/unwinder/sentry_unwinder.c",
-    # unix platform (non-Windows branch)
-    "src/sentry_random.c",
-    "src/sentry_unix_pageallocator.c",
-    "src/symbolizer/sentry_symbolizer_unix.c",
-    "src/path/sentry_path_unix.c",
-    "src/screenshot/sentry_screenshot_none.c",
-    # curl transport
-    "src/transports/sentry_transport_curl.c",
-    # inproc backend
-    "src/backends/sentry_backend_inproc.c",
-    # libbacktrace unwinder (glibc + macOS default)
-    "src/unwinder/sentry_unwinder_libbacktrace.c",
-    # vendored single-file deps compiled into libsentry (top CMakeLists
-    # add_library / target_sources); jsmn.h is header-only.
-    "vendor/mpack.c",
-    "vendor/stb_sprintf.c",
-]
-
-cc_library(
-    name = "sentry",
-    srcs = _SENTRY_SRCS + glob([
-        "src/**/*.h",
-        "vendor/*.h",
-    ]) + select({
-        "@platforms//os:macos": ["src/modulefinder/sentry_modulefinder_apple.c"],
-        "@platforms//os:linux": ["src/modulefinder/sentry_modulefinder_linux.c"],
-    }),
-    hdrs = ["include/sentry.h"],
-    includes = [
-        "include",
-        "src",
-    ],
-    linkstatic = True,
-    # The symbolizer (dladdr) + sync (pthread) call into libdl/libpthread; macOS
-    # folds both into libSystem. Declared here so the library is self-contained
-    # and consumers can't forget them (also declared on //src/metadata for the
-    # standalone metadata cc_tests).
-    linkopts = select({
-        "@platforms//os:linux": [
-            "-ldl",
-            "-lpthread",
-        ],
-        "//conditions:default": [],
-    }),
-    local_defines = [
-        "SENTRY_BACKEND_INPROC",
-        "SENTRY_WITH_UNWINDER_LIBBACKTRACE",
-        "SENTRY_HANDLER_STACK_SIZE=64",
-        "SIZEOF_LONG=8",  # LP64 (all three SIPI targets); sentry_sync.h reads it
-    ],
-    deps = [
-        "@curl",
-        "@openssl//:crypto",
-        "@openssl//:ssl",
-    ],
-)
