+++ src/SipiHttpServer.cpp
@@ -48,7 +48,6 @@
 #include "SipiPeakMemory.h"
 #include "SipiRateLimiter.h"
 #include "iiifparser/SipiDecodeDims.h"
-#include "observability/sentry.h"
 #include "favicon.h"
 #include "ffi/engine_context.h"
 #include "ffi/preflight.h"
@@ -62,15 +61,7 @@ using namespace shttps;
 
 namespace Sipi {
 
-// Hoist the observability:: free functions + types into Sipi:: scope inside
-// this TU so the existing call-site shape (`capture_image_error`, `ImageContext`,
-// etc.) keeps working without per-call qualification.
-using observability::capture_image_error;
-using observability::get_file_size;
-using observability::ImageContext;
 using observability::Metrics;
-using observability::populate_from_image;
-using observability::SipiMode;
 
 /*!
  * The name of the Lua function that checks permissions before a file is returned to an HTTP client.
