--- src/observability/sentry.cpp
@@ -1,124 +1,0 @@
-/*
- * Copyright © 2016 - 2024 Swiss National Data and Service Center for the Humanities and/or DaSCH Service Platform
- * contributors. SPDX-License-Identifier: AGPL-3.0-or-later
- */
-
-#include "observability/sentry.h"
-
-#include <sys/stat.h>
-
-namespace Sipi::observability {
-
-size_t get_file_size(const std::string &path)
-{
-  struct stat st{};
-  if (stat(path.c_str(), &st) == 0) {
-    return static_cast<size_t>(st.st_size);
-  }
-  return 0;
-}
-
-// `format_type_to_string` lives with the format enum in
-// src/iiifparser/SipiQualityFormat.{h,cpp}: this header (and so this package)
-// no longer has to depend on the iiifparser type just to stringify a format.
-
-// `populate_from_image` lives in src/populate_from_image.cpp under //src:engine
-// because it dereferences `Sipi::SipiImage`, which belongs to that target.
-// Co-locating it here would force //src/observability to depend on the engine,
-// creating a cycle (the engine already depends on observability for this
-// surface). Move the implementation back into this TU once SipiImage gets its
-// own package (DEV-6388 / DEV-6395).
-
-void capture_image_error(const std::string &error_message,
-  const std::string &phase,
-  const ImageContext &ctx,
-  SipiMode mode,
-  sentry_level_e level)
-{
-  const char *mode_str = (mode == SipiMode::CLI) ? "cli" : "server";
-
-  sentry_value_t event = sentry_value_new_event();
-  sentry_value_set_by_key(event,
-    "level",
-    sentry_value_new_string(level == SENTRY_LEVEL_ERROR  ? "error"
-                            : level == SENTRY_LEVEL_WARNING ? "warning"
-                                                            : "fatal"));
-
-  sentry_value_t message = sentry_value_new_object();
-  sentry_value_set_by_key(message, "formatted", sentry_value_new_string(error_message.c_str()));
-  sentry_value_set_by_key(event, "message", message);
-
-  // Set searchable tags directly on the event (not the global scope) for thread safety.
-  sentry_value_t tags = sentry_value_new_object();
-  sentry_value_set_by_key(tags, "sipi.mode", sentry_value_new_string(mode_str));
-  sentry_value_set_by_key(tags, "sipi.phase", sentry_value_new_string(phase.c_str()));
-  if (!ctx.output_format.empty()) {
-    sentry_value_set_by_key(tags, "sipi.output_format", sentry_value_new_string(ctx.output_format.c_str()));
-  }
-  if (!ctx.colorspace.empty()) {
-    sentry_value_set_by_key(tags, "sipi.colorspace", sentry_value_new_string(ctx.colorspace.c_str()));
-  }
-  if (ctx.bps > 0) {
-    sentry_value_set_by_key(tags, "sipi.bps", sentry_value_new_string(std::to_string(ctx.bps).c_str()));
-  }
-  if (!ctx.request_uri.empty()) {
-    sentry_value_set_by_key(tags, "sipi.request_uri", sentry_value_new_string(ctx.request_uri.c_str()));
-  }
-  sentry_value_set_by_key(event, "tags", tags);
-
-  // Build "Image" context with full details
-  sentry_value_t image_ctx = sentry_value_new_object();
-  sentry_value_set_by_key(image_ctx, "type", sentry_value_new_string("image"));
-
-  sentry_value_t image_data = sentry_value_new_object();
-  if (!ctx.input_file.empty()) {
-    sentry_value_set_by_key(image_data, "input_file", sentry_value_new_string(ctx.input_file.c_str()));
-  }
-  if (!ctx.output_file.empty()) {
-    sentry_value_set_by_key(image_data, "output_file", sentry_value_new_string(ctx.output_file.c_str()));
-  }
-  if (!ctx.output_format.empty()) {
-    sentry_value_set_by_key(image_data, "output_format", sentry_value_new_string(ctx.output_format.c_str()));
-  }
-  if (ctx.width > 0 || ctx.height > 0) {
-    sentry_value_set_by_key(image_data, "width", sentry_value_new_int32(static_cast<int32_t>(ctx.width)));
-    sentry_value_set_by_key(image_data, "height", sentry_value_new_int32(static_cast<int32_t>(ctx.height)));
-  }
-  if (ctx.channels > 0) {
-    sentry_value_set_by_key(image_data, "channels", sentry_value_new_int32(static_cast<int32_t>(ctx.channels)));
-  }
-  if (ctx.bps > 0) {
-    sentry_value_set_by_key(image_data, "bps", sentry_value_new_int32(static_cast<int32_t>(ctx.bps)));
-  }
-  if (!ctx.colorspace.empty()) {
-    sentry_value_set_by_key(image_data, "colorspace", sentry_value_new_string(ctx.colorspace.c_str()));
-  }
-  if (!ctx.icc_profile_type.empty()) {
-    sentry_value_set_by_key(image_data, "icc_profile_type", sentry_value_new_string(ctx.icc_profile_type.c_str()));
-  }
-  if (!ctx.orientation.empty()) {
-    sentry_value_set_by_key(image_data, "orientation", sentry_value_new_string(ctx.orientation.c_str()));
-  }
-  if (ctx.file_size_bytes > 0) {
-    sentry_value_set_by_key(
-      image_data, "file_size_bytes", sentry_value_new_double(static_cast<double>(ctx.file_size_bytes)));
-  }
-  if (!ctx.request_uri.empty()) {
-    sentry_value_set_by_key(image_data, "request_uri", sentry_value_new_string(ctx.request_uri.c_str()));
-  }
-
-  sentry_value_set_by_key(image_ctx, "data", image_data);
-
-  // Add the Image context to the event
-  sentry_value_t contexts = sentry_value_new_object();
-  sentry_value_set_by_key(contexts, "Image", image_ctx);
-  sentry_value_set_by_key(event, "contexts", contexts);
-
-  sentry_capture_event(event);
-
-  // In CLI mode, block briefly to ensure the event is sent before process exit.
-  // In server mode, use non-blocking flush to avoid stalling request threads.
-  sentry_flush(mode == SipiMode::CLI ? 2000 : 0);
-}
-
-}// namespace Sipi::observability
